Rendimiento mejorado y permite ejecución de hilos en paralelo
